@@ -2,8 +2,8 @@ package de.caritas.cob.consultingtypeservice.api.tenant;
 
 import static de.caritas.cob.consultingtypeservice.api.tenant.TenantResolver.TECHNICAL_TENANT_ID;
 
-import javax.persistence.EntityManager;
-import javax.persistence.PersistenceContext;
+import jakarta.persistence.EntityManager;
+import jakarta.persistence.PersistenceContext;
 import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
 import org.aspectj.lang.annotation.Aspect;
@@ -20,7 +20,6 @@ public class TenantAspect {
 
   @PersistenceContext private final @NonNull EntityManager entityManager;
 
-  // TODO refactor this package when the service name will be refactored
   @Before("execution(* de.caritas.cob.consultingtypeservice.api.port..*(..)))")
   public void beforeQueryAspect() {
 
