@@ -1,7 +1,7 @@
 package de.caritas.cob.consultingtypeservice.api.auth;
 
 import java.util.Collection;
-import java.util.Optional;
+import java.util.Objects;
 import java.util.Set;
 import java.util.stream.Collectors;
 import org.springframework.security.core.GrantedAuthority;
@@ -25,12 +25,11 @@ public class RoleAuthorizationAuthorityMapper implements GrantedAuthoritiesMappe
     return mapAuthorities(roleNames);
   }
 
-  private Set<GrantedAuthority> mapAuthorities(Set<String> roleNames) {
-    return roleNames.parallelStream()
-        .map(UserRole::getRoleByValue)
-        .filter(Optional::isPresent)
-        .map(Optional::get)
-        .map(Authority::getAuthoritiesByUserRole)
+  public Set<GrantedAuthority> mapAuthorities(Set<String> roleNames) {
+    return roleNames.stream()
+        .map(Authority::fromRoleName)
+        .filter(Objects::nonNull)
+        .map(Authority::getGrantedAuthorities)
         .flatMap(Collection::parallelStream)
         .map(SimpleGrantedAuthority::new)
         .collect(Collectors.toSet());
