@@ -5,14 +5,14 @@ import de.caritas.cob.consultingtypeservice.api.exception.UnexpectedErrorExcepti
 import de.caritas.cob.consultingtypeservice.api.service.LogService;
 import de.caritas.cob.consultingtypeservice.api.tenant.TenantContext;
 import de.caritas.cob.consultingtypeservice.schemas.model.ConsultingType;
+import jakarta.annotation.PostConstruct;
 import java.io.File;
 import java.io.IOException;
 import java.net.MalformedURLException;
 import java.net.URISyntaxException;
 import java.nio.file.InvalidPathException;
-import java.nio.file.Paths;
+import java.nio.file.Path;
 import java.util.stream.Stream;
-import javax.annotation.PostConstruct;
 import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
 import org.apache.commons.lang3.ArrayUtils;
@@ -62,7 +62,7 @@ public class ConsultingTypeLoader {
 
   private File[] determineConsultingTypeConfigurationFiles() {
     try {
-      var dirUrl = Paths.get(consultingTypesFilePath).toUri().toURL();
+      var dirUrl = Path.of(consultingTypesFilePath).toUri().toURL();
       return new File(dirUrl.toURI())
           .listFiles((dir, name) -> name.toLowerCase().endsWith(".json"));
     } catch (URISyntaxException | MalformedURLException | InvalidPathException exception) {
