@@ -47,7 +47,7 @@ public class ConsultingTypeMongoRepositoryService implements ConsultingTypeRepos
 
     if (byId.isEmpty()) {
       throw new NotFoundException(
-          String.format("Consulting type with id %s not found.", consultingTypeId));
+          "Consulting type with id %s not found.".formatted(consultingTypeId));
     }
     return byId.get();
   }
@@ -62,9 +62,7 @@ public class ConsultingTypeMongoRepositoryService implements ConsultingTypeRepos
     return consultingTypeRepository.findBySlug(slug).stream()
         .findFirst()
         .orElseThrow(
-            () ->
-                new NotFoundException(
-                    String.format("Consulting type with slug %s not found.", slug)));
+            () -> new NotFoundException("Consulting type with slug %s not found.".formatted(slug)));
   }
 
   @Override
@@ -86,9 +84,8 @@ public class ConsultingTypeMongoRepositoryService implements ConsultingTypeRepos
     log.debug("Using tenant unaware repository service to try to add a consulting type");
     if (isConsultingTypeWithGivenIdPresent(consultingType) || slugIsNotValid(consultingType)) {
       LogService.logWarning(
-          String.format(
-              "Could not add consulting type. id %s or slug %s is not unique",
-              consultingType.getId(), consultingType.getSlug()));
+          "Could not add consulting type. id %s or slug %s is not unique"
+              .formatted(consultingType.getId(), consultingType.getSlug()));
       return Optional.empty();
     } else {
       final var consultingTypeEntity = new ConsultingTypeEntity();
