@@ -1,9 +1,11 @@
 package de.caritas.cob.consultingtypeservice.api.exception;
 
+import java.io.Serial;
+
 /** Exception, when unexpected error occurs. */
 public class UnexpectedErrorException extends RuntimeException {
 
-  private static final long serialVersionUID = -5083156826149548581L;
+  @Serial private static final long serialVersionUID = -5083156826149548581L;
 
   /** Unexpected error exception. */
   public UnexpectedErrorException() {
