@@ -1,10 +1,11 @@
 package de.caritas.cob.consultingtypeservice.api.exception.httpresponses;
 
 import de.caritas.cob.consultingtypeservice.api.service.LogService;
+import java.io.Serial;
 
 public class BadRequestException extends CustomHttpStatusException {
 
-  private static final long serialVersionUID = -8047408802295905803L;
+  @Serial private static final long serialVersionUID = -8047408802295905803L;
 
   /**
    * BadRequest exception.
